# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /study/auth/signup:
    post:
      operationId: studySignup
      tags: [study-auth]
      summary: Sign up as an independent learner
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudyUserSignup"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudyAuthResult"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudyError"
        "409":
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudyError"
  /study/auth/login:
    post:
      operationId: studyLogin
      tags: [study-auth]
      summary: Log in
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudyUserLogin"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudyAuthResult"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudyError"
  /study/auth/me:
    get:
      operationId: studyGetMe
      tags: [study-auth]
      summary: Get current learner
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudyUser"
        "401":
          description: Not signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudyError"
...
  /study/exams:
    get:
      operationId: listStudyExams
      tags: [study-exams]
      summary: List mock exams
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StudyMockExam"
    post:
      operationId: createStudyExam
      tags: [study-exams]
      summary: Generate a mock exam
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudyMockExamInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudyMockExam"
  /study/exams/{examId}:
    get:
      operationId: getStudyExam
      tags: [study-exams]
      summary: Get exam details
      parameters:
        - name: examId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudyMockExam"
  /study/exams/{examId}/submit:
    post:
      operationId: submitStudyExam
      tags: [study-exams]
      summary: Submit completed exam for grading
      parameters:
        - name: examId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudyMockExamSubmit"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudyMockExamResult"

  # ─── Study Tutor ───
...
components:
  schemas:
...
    StudyError:
      type: object
      properties:
        error:
          type: string
        details:
          type: object
      required:
        - error
...
    StudyUser:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        name:
          type: string
        subscriptionStatus:
          type: string
        subscriptionCurrentPeriodEnd:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - email
        - name
        - subscriptionStatus
        - createdAt
    StudyUserSignup:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        name:
          type: string
          minLength: 1
      required:
        - email
        - password
        - name
    StudyUserLogin:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required:
        - email
        - password
    StudyAuthResult:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/StudyUser"
        message:
          type: string
      required:
        - user
...
    StudyMockExam:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        title:
          type: string
        materialId:
          type: ["string", "null"]
        questionCount:
          type: integer
        timeLimitMinutes:
          type: integer
        status:
          type: string
        questions:
          type: array
          items:
            $ref: "#/components/schemas/StudyMockExamQuestion"
        score:
          type: ["number", "null"]
        maxScore:
          type: ["integer", "null"]
        completedAt:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - userId
        - title
        - questionCount
        - timeLimitMinutes
        - status
        - questions
        - createdAt
    StudyMockExamQuestion:
      type: object
      properties:
        id:
          type: string
        prompt:
          type: string
        options:
          type: array
          items:
            type: string
        correctOptionIndex:
          type: integer
        explanation:
          type: string
        conceptId:
          type: ["string", "null"]
        points:
          type: integer
      required:
        - id
        - prompt
        - options
        - correctOptionIndex
        - explanation
        - points
    StudyMockExamInput:
      type: object
      properties:
        title:
          type: string
        materialId:
          type: ["string", "null"]
        conceptIds:
          type: array
          items:
            type: string
        questionCount:
          type: integer
          default: 20
        timeLimitMinutes:
          type: integer
          default: 30
      required: []
    StudyMockExamSubmit:
      type: object
      properties:
        answers:
          type: array
          items:
            type: object
            properties:
              questionId:
                type: string
              selectedOptionIndex:
                type: integer
            required:
              - questionId
              - selectedOptionIndex
        timeSpentSeconds:
          type: integer
      required:
        - answers
        - timeSpentSeconds
    StudyMockExamResult:
      type: object
      properties:
        score:
          type: number
        maxScore:
          type: integer
        percentage:
          type: number
        correctCount:
          type: integer
        timeSpentSeconds:
          type: integer
        conceptBreakdown:
          type: array
          items:
            $ref: "#/components/schemas/StudyConceptBreakdown"
      required:
        - score
        - maxScore
        - percentage
        - correctCount
        - timeSpentSeconds
        - conceptBreakdown
    StudyConceptBreakdown:
      type: object
      properties:
        conceptId:
          type: string
        conceptTitle:
          type: string
        correct:
          type: integer
        total:
          type: integer
        percentage:
          type: number
      required:
        - conceptId
        - conceptTitle
        - correct
        - total
        - percentage
...