# lib/api-zod/src/generated/types/studyMockExamInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface StudyMockExamInput {
  title?: string;
  /** @nullable */
  materialId?: string | null;
  conceptIds?: string[];
  questionCount?: number;
  timeLimitMinutes?: number;
}