# lib/api-zod/src/generated/types/studyMockExamQuestion.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface StudyMockExamQuestion {
  id: string;
  prompt: string;
  options: string[];
  correctOptionIndex: number;
  explanation: string;
  /** @nullable */
  conceptId?: string | null;
  points: number;
}